enterprise_ai/requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
chromadb>=0.4.0
networkx>=3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
prometheus-client>=0.19.0
PyPDF2>=3.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0
slack-sdk>=3.23.0
notion-client>=2.2.0
grafana-api>=1.0.3
httpx>=0.25.0
tiktoken>=0.5.0
numpy>=1.24.0
